+++ lib/index.ts
@@ -9,6 +9,11 @@ export { Contact } from './contact'
 export { AccountManager as DeltaChat }
 export { Locations } from './locations'
 export { Lot } from './lot'
-export { Message, MessageState, MessageViewType } from './message'
+export {
+  Message,
+  MessageState,
+  MessageViewType,
+  MessageDownloadState,
+} from './message'
 
 export * from './types'